[ALPS05420914] CCU:fix android 32/64bit
Fix struct CcuMemInfo & ccu_control_info for Android 32/64-bit.

MTK-Commit-Id: 639371f9fd44eb8486a9f3ebb522d3f0efc8964c

CR-Id: ALPS05420914
Feature: Camera Control Unit(ccu)

+++ drivers/misc/mediatek/ccu/src/mt6833/ccu_drv.c
@@ -652,8 +652,7 @@ static long ccu_ioctl(struct file *flip, unsigned int cmd,
 			break;
 		}
 
-		ret = copy_from_user(indata,
-			(void *)((unsigned long long)(msg.inDataPtr)), msg.inDataSize);
+		ret = copy_from_user(indata, (void *)msg.inDataPtr, msg.inDataSize);
 		if (ret != 0) {
 			LOG_ERR(
 			"CCU_IOCTL_IPC_SEND_CMD copy_to_user 2 failed: %d\n",
@@ -667,8 +666,7 @@ static long ccu_ioctl(struct file *flip, unsigned int cmd,
 		(enum IMGSENSOR_SENSOR_IDX)msg.sensor_idx,
 		msg.msg_id, indata, msg.inDataSize, outdata, msg.outDataSize);
 
-		ret = copy_to_user((void *)((unsigned long long)(msg.outDataPtr)),
-				    outdata, msg.outDataSize);
+		ret = copy_to_user((void *)msg.outDataPtr, outdata, msg.outDataSize);
 		kfree(indata);
 		kfree(outdata);
 		break;
